+++ components/subscriptions/subscriptionsTabScreen/index.tsx
@@ -1,7 +1,7 @@
 import { View, Text } from 'react-native';
 import React from 'react';
-import CalendarPicker from './CalendarPicker';
-import SubscriptionsList from './SubscriptionsList';
+import CalendarPicker from './calendar-picker';
+import SubscriptionsList from './subscriptions-list';
 
 const SubscriptionTabScreen = () => {
   return (
